testing_azure/requirements.txt
streamlit
azure-storage-blob
python-dotenv
